pyproject.toml
[project]
name = "hogwild-uxr"
version = "4.0.0"
description = "UX Research Sense-Making Pipeline — Orchestrator MCP + Copilot CLI Skills"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "Jake Rhodes"}
]
dependencies = [
    "mcp>=1.0.0",
    "python-docx>=1.1.0",
    "pyyaml>=6.0",
    "jsonschema>=4.20.0",
]

[project.scripts]
hogwild-uxr = "hogwild_uxr.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hogwild_uxr"]
